bootstrap the koans Angular app
For a start AngularJS is loaded from a CDN. The Koans application is
implemented in the 'koans' Angular module defined in 'app.js'. This
application sets the application name.

While Angular is loading the entire body of the app is hidden using the
ng-cloak directive. Angular automatically removes this directive causing
the content to show only after Angular is loaded.