greet who enters her name
Show a personal greeting message if someone enters her name. Otherwise a
default message is shown.

The text input is bound to the 'name' model object using the ng-model
directive. Because the input is bound to the same model value that is
rendered in the greeting message the greeting message updates when the
input changes.